+++ crates/cheatnet/src/predeployment/mod.rs
@@ -1,1 +1,2 @@
-pub mod strk;
+pub mod erc20;
+pub mod predeployed_contract;