+++ crates/sncast/src/starknet_commands/account/deploy.rs
@@ -7,7 +7,8 @@ use sncast::helpers::braavos::BraavosAccountFactory;
 use sncast::helpers::constants::{BRAAVOS_BASE_ACCOUNT_CLASS_HASH, KEYSTORE_PASSWORD_ENV_VAR};
 use sncast::helpers::fee::{FeeArgs, FeeSettings};
 use sncast::helpers::rpc::RpcArgs;
-use sncast::response::structs::InvokeResponse;
+use sncast::response::account::deploy::AccountDeployResponse;
+use sncast::response::invoke::InvokeResponse;
 use sncast::{
     AccountType, WaitForTx, apply_optional_fields, chain_id_to_network_name,
     check_account_file_exists, get_account_data_from_accounts_file, get_account_data_from_keystore,
@@ -52,7 +53,7 @@ pub async fn deploy(
     account: &str,
     keystore_path: Option<Utf8PathBuf>,
     fee_args: FeeArgs,
-) -> Result<InvokeResponse> {
+) -> Result<AccountDeployResponse> {
     if let Some(keystore_path_) = keystore_path {
         deploy_from_keystore(
             provider,
@@ -63,6 +64,7 @@ pub async fn deploy(
             keystore_path_,
         )
         .await
+        .map(Into::into)
     } else {
         let account_name = deploy_args
             .name
@@ -78,6 +80,7 @@ pub async fn deploy(
             wait_config,
         )
         .await
+        .map(Into::into)
     }
 }